+++ framework/tests/ut/passes/src/test_health_report.cpp
@@ -79,5 +79,5 @@ TEST_F(HealthReportTest, TestDD) {
     TestLoopViewAssembleCopy(t0, t1, blockTable, out, s);
 
     auto funcMap = Program::GetInstance().GetFunctionMap();
-    APASS_LOG_INFO_F(Elements::Function, "%d", funcMap.size());
+    APASS_LOG_INFO_F(Elements::Function, "%d", static_cast<int>(funcMap.size()));
 }
